requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
coreschema==0.0.4
Django==4.2.13
django-js-asset==2.2.0
django-mptt==0.14.0
djangorestframework==3.15.1
idna==3.7
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pillow==10.3.0
requests==2.31.0
sqlparse==0.5.0
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
